workspaces/weather_app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
atomic-agents==2.7.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
fastuuid==0.14.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.1
identify==2.6.17
idna==3.11
importlib_metadata==8.7.1
instructor==1.13.0
Jinja2==3.1.6
jiter==0.11.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
linkify-it-py==2.1.0
litellm==1.82.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
multidict==6.7.1
nodeenv==1.10.0
openai==2.28.0
packaging==26.0
platformdirs==4.9.4
pre_commit==4.5.1
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyfiglet==1.0.4
Pygments==2.19.2
PyJWT==2.12.1
python-discovery==1.1.3
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==13.9.4
rpds-py==0.30.0
shellingham==1.5.4
smmap==5.0.3
sniffio==1.3.1
sse-starlette==3.3.2
starlette==0.52.1
tenacity==9.1.4
textual==5.3.0
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
ty==0.0.23
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uc-micro-py==2.0.0
urllib3==2.6.3
uvicorn==0.41.0
virtualenv==21.2.0
watchdog==6.0.0
yarl==1.23.0
zipp==3.23.0
groq
